+++ ArticleTypes.ts
@@ -1,0 +1,3 @@
+export enum _ARTICLE_TYPES {
+	ARTICLE = 'article',
+}
