--- studio/src/schemas/documents/articles/types.ts
@@ -1,4 +1,0 @@
-export enum _ARTICLE_TYPES {
-	article = 'article',
-	test_article = 'test_article',
-}
