+++ studio/src/schemas/documents/articles/utils/baseArticle.tsx
@@ -1,0 +1,13 @@
+import { _ARTICLE_FIELDS } from "@schemas.studio/documents/articles/utils/fields";
+import { _ARTICLE_PREVIEW } from "@schemas.studio/documents/articles/utils/preview";
+import { defineType } from "sanity";
+
+export const _BASE_ARTICLE = defineType({
+	title: 'Base Article',
+	name: 'baseArticle',
+	type: 'document',
+	fields: [
+		..._ARTICLE_FIELDS
+	],
+	preview: _ARTICLE_PREVIEW
+})
