+++ studio/src/schemas/documents/articles/utils/groups.tsx
@@ -1,0 +1,9 @@
+import { FieldGroupDefinition } from "sanity";
+
+const _ARTICLE_GROUPS: FieldGroupDefinition[] = [
+	{
+		name: 'about',
+		title: 'About',
+		// default: true,
+	},
+]
