+++ studio/src/schemas/documents/pages/blocks/FeaturedTaxonomies.tsx
@@ -1,4 +1,4 @@
-import constructors from '../constructors';
+import constructors from '../utils';
 import { defineField } from 'sanity';
 
 import { SwatchIcon } from '@heroicons/react/24/solid'
