+++ studio/src/schemas/documents/pages/blocks/Gallery.tsx
@@ -1,5 +1,5 @@
 import { defineField } from 'sanity';
-import constructors from '../constructors';
+import constructors from '../utils';
 import { customMediaAssetSource } from '../../../../lib/assetSource';
 
 import { Squares2X2Icon } from '@heroicons/react/24/solid'
