+++ studio/src/schemas/documents/pages/utils/preview.tsx
@@ -1,0 +1,15 @@
+import { camelCaseToWords } from "@lib.studio/stringFunctions"
+
+export const _BLOCK_PREVIEW = {
+		select: {
+			type: '_type',
+			title: 'title',
+			// logo: 'logo',
+		},
+		prepare(value: any) {
+			const { type, title } = value
+			return {
+				title: type ? camelCaseToWords(type) : 'Unknown Block Type',
+			}
+		},
+	}
